fix: not propagating synchronization back to AAI while logging
In normal circumstances, if user is added/removed from a role inside a community, this change is propagated to AAI. This is not required during login, as these changes are the result of applying AAI settings on the repository. This change adds a meaning of disabling the propagation and applies it to the logging process.